Added initramfs.runsize=24M to Linux kernel bootargs
The default value is 10% of physical RAM size.
On a board with less then 256MB RAM, 10% is not enough for systemd to restart a service:
Failed to reload daemon: Refusing to reload, not enough space available on /run/systemd.
Currently, 13.2M are free, but a safety buffer of 16.0M is enforced.