Update mk-sd-card
Sometimes an SD card reader might have two physical slots that support SD and microSD in a single adapter, but the adapter creates two mass storage devices. If unluckily the SD slot is in front of the microSD slot, the script will try to read the empty medium and fail.